+++ src/main/scala/licos/json/validation/lobby/IdSearchValidation.scala
@@ -5,5 +5,5 @@ import play.api.libs.json.Reads.{max, min}
 import play.api.libs.functional.syntax._
 
 object IdSearchValidation {
-  val idForSearching: Reads[Int] = min(-1) keepAnd max(Int.MaxValue)
+  val idForSearching: Reads[Long] = min(-1L) keepAnd max(Long.MaxValue)
 }
